registration specific functions moved from Util:: to Registration::
