+++ app/mon/mon_cli/src/ecal_mon_cli.cpp
@@ -353,7 +353,7 @@ void ProcProto(const std::string& topic_name, int msg_count)
 
   // get topic type
   eCAL::SDataTypeInformation topic_info;
-  eCAL::Util::GetTopicDataTypeInformation(topic_name, topic_info);
+  eCAL::Registration::GetTopicDataTypeInformation(topic_name, topic_info);
   if(topic_info.name.empty())
   {
     std::cout << "could not get type name for topic " << topic_name << std::endl;
