+++ app/mon/mon_gui/src/widgets/models/topic_tree_item.cpp
@@ -95,7 +95,7 @@ QVariant TopicTreeItem::data(Columns column, Qt::ItemDataRole role) const
         if (!monitor_topic_name.empty())
         {
           eCAL::SDataTypeInformation topic_info;
-          eCAL::Util::GetTopicDataTypeInformation(monitor_topic_name, topic_info);
+          eCAL::Registration::GetTopicDataTypeInformation(monitor_topic_name, topic_info);
           return topic_info.encoding.c_str();
         }
       }
@@ -115,7 +115,7 @@ QVariant TopicTreeItem::data(Columns column, Qt::ItemDataRole role) const
         if (!monitor_topic_name.empty())
         {
           eCAL::SDataTypeInformation topic_info;
-          eCAL::Util::GetTopicDataTypeInformation(monitor_topic_name, topic_info);
+          eCAL::Registration::GetTopicDataTypeInformation(monitor_topic_name, topic_info);
           return topic_info.name.c_str();
         }
       }
