+++ ecal/core/include/ecal/cimpl/ecal_util_cimpl.h
@@ -1,6 +1,6 @@
 /* ========================= eCAL LICENSE =================================
  *
- * Copyright (C) 2016 - 2019 Continental Corporation
+ * Copyright (C) 2016 - 2024 Continental Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,7 +26,6 @@
 #define ecal_util_cimpl_h_included
 
 #include <ecal/ecal_os.h>
-#include <ecal/ecalc_types.h>
 
 #ifdef __cplusplus
 extern "C"
@@ -64,144 +63,6 @@ extern "C"
    * @param state_  Switch on message loop back.. 
   **/
   ECALC_API void eCAL_Util_EnableLoopback(int state_);
-
-  /**
-   * @brief Gets type name of the specified topic. 
-   *
-   * @param       topic_name_      Topic name. 
-   * @param [out] topic_type_      Pointer to store the type name information. 
-   * @param       topic_type_len_  Length of allocated buffer or ECAL_ALLOCATE_4ME if
-   *                               eCAL should allocate the buffer for you (see eCAL_FreeMem). 
-   *
-   * @return  Type name buffer length or zero if failed. 
-  **/
-  ECALC_API int eCAL_Util_GetTopicTypeName(const char* topic_name_, void* topic_type_, int topic_type_len_);
-
-  /**
-   * @brief Gets encoding of the specified topic.
-   *
-   * @param       topic_name_           Topic name.
-   * @param [out] topic_encoding_       Pointer to store the encoding information.
-   * @param       topic_encoding__len_  Length of allocated buffer or ECAL_ALLOCATE_4ME if
-   *                                    eCAL should allocate the buffer for you (see eCAL_FreeMem).
-   *
-   * @return  Type name buffer length or zero if failed.
-  **/
-  ECALC_API int eCAL_Util_GetTopicEncoding(const char* topic_name_, void* topic_encoding_, int topic_encoding_len_);
-
-  /**
-   * @brief Gets type description of the specified topic. 
-   *
-   * @param       topic_name_      Topic name. 
-   * @param [out] topic_desc_      Pointer to store the type description0 information. 
-   * @param       topic_desc_len_  Length of allocated buffer or ECAL_ALLOCATE_4ME if
-   *                               eCAL should allocate the buffer for you (see eCAL_FreeMem). 
-   *
-   * @return  Type name buffer length or zero if failed. 
-  **/
-  ECALC_API int eCAL_Util_GetTopicDescription(const char* topic_name_, void* topic_desc_, int topic_desc_len_);
-
-  /**
-   * @brief Gets service method request type name.
-   *
-   * @param       service_name_   Service name. 
-   * @param       method_name_    Method name.
-   * @param [out] req_type_       Pointer to store the request type.
-   * @param       req_type_len_   Length of allocated buffer or ECAL_ALLOCATE_4ME if
-   *                              eCAL should allocate the buffer for you (see eCAL_FreeMem). 
-   *
-   * @return  Type name buffer length or zero if failed.
-  **/
-  ECALC_API int eCAL_Util_GetServiceRequestTypeName(const char* service_name_, const char* method_name_, void* req_type_, int req_type_len_);
-
-  /**
-   * @brief Gets service method response type name.
-   *
-   * @param       service_name_   Service name.
-   * @param       method_name_    Method name.
-   * @param [out] resp_type_      Pointer to store the response type.
-   * @param       resp_type_len_  Length of allocated buffer or ECAL_ALLOCATE_4ME if
-   *
-   * @return  Type name buffer length or zero if failed.
-  **/
-  ECALC_API int eCAL_Util_GetServiceResponseTypeName(const char* service_name_, const char* method_name_, void* resp_type_, int resp_type_len_);
-
-  /**
-   * @brief Gets service method request description.
-   *
-   * @param       service_name_   Service name.
-   * @param       method_name_    Method name.
-   * @param [out] req_desc_       Pointer to store the request description.
-   * @param       req_desc_len_   Length of allocated buffer or ECAL_ALLOCATE_4ME if
-   *                              eCAL should allocate the buffer for you (see eCAL_FreeMem).
-   *
-   * @return  Request description buffer length or zero if failed.
-  **/
-  ECALC_API int eCAL_Util_GetServiceRequestDescription(const char* service_name_, const char* method_name_, void* req_desc_, int req_desc_len_);
-
-  /**
-   * @brief Gets service method response description.
-   *
-   * @param       service_name_   Service name.
-   * @param       method_name_    Method name.
-   * @param [out] resp_desc_      Pointer to store the response description.
-   * @param       resp_desc_len_  Length of allocated buffer or ECAL_ALLOCATE_4ME if
-   *                              eCAL should allocate the buffer for you (see eCAL_FreeMem).
-   *
-   * @return  Response description buffer length or zero if failed.
-  **/
-  ECALC_API int eCAL_Util_GetServiceResponseDescription(const char* service_name_, const char* method_name_, void* resp_desc_, int resp_desc_len_);
-
-  /**
-   * @brief Gets client method request type name.
-   *
-   * @param       client_name_   Client name.
-   * @param       method_name_   Method name.
-   * @param [out] req_type_      Pointer to store the request type.
-   * @param       req_type_len_  Length of allocated buffer or ECAL_ALLOCATE_4ME if
-   *                             eCAL should allocate the buffer for you (see eCAL_FreeMem).
-   *
-   * @return  Type name buffer length or zero if failed.
-  **/
-  ECALC_API int eCAL_Util_GetClientRequestTypeName(const char* client_name_, const char* method_name_, void* req_type_, int req_type_len_);
-
-  /**
-   * @brief Gets client method response type name.
-   *
-   * @param       client_name_   Client name.
-   * @param       method_name_   Method name.
-   * @param [out] resp_type_     Pointer to store the response type.
-   * @param       resp_type_len_ Length of allocated buffer or ECAL_ALLOCATE_4ME if
-   *
-   * @return  Type name buffer length or zero if failed.
-  **/
-  ECALC_API int eCAL_Util_GetClientResponseTypeName(const char* client_name_, const char* method_name_, void* resp_type_, int resp_type_len_);
-
-  /**
-   * @brief Gets client method request description.
-   *
-   * @param       client_name_   Client name.
-   * @param       method_name_   Method name.
-   * @param [out] req_desc_      Pointer to store the request description.
-   * @param       req_desc_len_  Length of allocated buffer or ECAL_ALLOCATE_4ME if
-   *                             eCAL should allocate the buffer for you (see eCAL_FreeMem).
-   *
-   * @return  Request description buffer length or zero if failed.
-  **/
-  ECALC_API int eCAL_Util_GetClientRequestDescription(const char* client_name_, const char* method_name_, void* req_desc_, int req_desc_len_);
-
-  /**
-   * @brief Gets client method response description.
-   *
-   * @param       client_name_   Client name.
-   * @param       method_name_   Method name.
-   * @param [out] resp_desc_     Pointer to store the response description.
-   * @param       resp_desc_len_ Length of allocated buffer or ECAL_ALLOCATE_4ME if
-   *                             eCAL should allocate the buffer for you (see eCAL_FreeMem).
-   *
-   * @return  Response description buffer length or zero if failed.
-  **/
-  ECALC_API int eCAL_Util_GetClientResponseDescription(const char* client_name_, const char* method_name_, void* resp_desc_, int resp_desc_len_);
 #ifdef __cplusplus
 }
 #endif /*__cplusplus*/
