+++ ecal/core/include/ecal/ecal.h
@@ -35,6 +35,7 @@
 #include <ecal/ecal_process.h>
 #include <ecal/ecal_monitoring.h>
 #include <ecal/ecal_publisher.h>
+#include <ecal/ecal_registration.h>
 #include <ecal/ecal_server.h>
 #include <ecal/ecal_service_info.h>
 #include <ecal/ecal_subscriber.h>
