+++ ecal/samples/cpp/services/ping_client_dyn/src/ping_client_dyn.cpp
@@ -49,7 +49,7 @@ int main(int argc, char **argv)
   // get service method type names
   std::string req_type;
   std::string resp_type;
-  if (!eCAL::Util::GetServiceTypeNames(service_name, method_name, req_type, resp_type))
+  if (!eCAL::Registration::GetServiceTypeNames(service_name, method_name, req_type, resp_type))
   {
     throw std::runtime_error("Could not get service type names !");
   }
@@ -57,7 +57,7 @@ int main(int argc, char **argv)
   // get service method type descriptions
   std::string req_desc;
   std::string resp_desc;
-  if (!eCAL::Util::GetServiceDescription(service_name, method_name, req_desc, resp_desc))
+  if (!eCAL::Registration::GetServiceDescription(service_name, method_name, req_desc, resp_desc))
   {
     throw std::runtime_error("Could not get service type descriptions !");
   }
