+++ ecal/tests/CMakeLists.txt
@@ -33,13 +33,14 @@ add_subdirectory(cpp/descgate_test)
 add_subdirectory(cpp/config_test)
 
 if(ECAL_CORE_REGISTRATION)
-  add_subdirectory(cpp/util_test)
+  add_subdirectory(cpp/registration_test)
 endif()
 
 add_subdirectory(cpp/event_test)
 add_subdirectory(cpp/expmap_test)
 add_subdirectory(cpp/serialization_test)
 add_subdirectory(cpp/topic2mcast_test)
+add_subdirectory(cpp/util_test)
 
 if(ECAL_CORE_REGISTRATION_SHM OR ECAL_CORE_TRANSPORT_SHM)
   add_subdirectory(cpp/io_memfile_test)
