+++ ecal/tests/cpp/registration_test/src/registration_getclients.cpp
@@ -26,15 +26,15 @@ enum {
   CMN_REGISTRATION_REFRESH_MS = (1000)
 };
 
-TEST(core_cpp_util, ClientExpiration)
+TEST(core_cpp_registration, ClientExpiration)
 {
   // initialize eCAL API
-  eCAL::Initialize(0, nullptr, "core_cpp_util");
+  eCAL::Initialize(0, nullptr, "core_cpp_registration");
 
   // enable loop back communication in the same process
   eCAL::Util::EnableLoopback(true);
 
-  std::map<eCAL::Util::SServiceMethod, eCAL::SServiceMethodInformation> client_info_map;
+  std::map<eCAL::Registration::SServiceMethod, eCAL::SServiceMethodInformation> client_info_map;
 
   // create simple client and let it expire
   {
@@ -50,14 +50,14 @@ TEST(core_cpp_util, ClientExpiration)
     eCAL::Process::SleepMS(2 * CMN_REGISTRATION_REFRESH_MS);
 
     // get all clients
-    eCAL::Util::GetClients(client_info_map);
+    eCAL::Registration::GetClients(client_info_map);
 
     // check size
     EXPECT_EQ(client_info_map.size(), 1);
 
     // check client/method names
-    std::set<eCAL::Util::SServiceMethod> client_method_names;
-    eCAL::Util::GetClientMethodNames(client_method_names);
+    std::set<eCAL::Registration::SServiceMethod> client_method_names;
+    eCAL::Registration::GetClientMethodNames(client_method_names);
     EXPECT_EQ(client_method_names.size(), 1);
     for (const auto& name : client_method_names)
     {
@@ -69,7 +69,7 @@ TEST(core_cpp_util, ClientExpiration)
     eCAL::Process::SleepMS(CMN_MONITORING_TIMEOUT_MS);
 
     // get all clients again, client should not be expired
-    eCAL::Util::GetClients(client_info_map);
+    eCAL::Registration::GetClients(client_info_map);
 
     // check size
     EXPECT_EQ(client_info_map.size(), 1);
@@ -80,7 +80,7 @@ TEST(core_cpp_util, ClientExpiration)
 
   // get all clients again, all clients 
   // should be removed from the map
-  eCAL::Util::GetClients(client_info_map);
+  eCAL::Registration::GetClients(client_info_map);
 
   // check size
   EXPECT_EQ(client_info_map.size(), 0);
@@ -89,15 +89,15 @@ TEST(core_cpp_util, ClientExpiration)
   eCAL::Finalize();
 }
 
-TEST(core_cpp_util, ClientEqualQualities)
+TEST(core_cpp_registration, ClientEqualQualities)
 {
   // initialize eCAL API
-  eCAL::Initialize(0, nullptr, "core_cpp_util");
+  eCAL::Initialize(0, nullptr, "core_cpp_registration");
 
   // enable loop back communication in the same process
   eCAL::Util::EnableLoopback(true);
 
-  std::map<eCAL::Util::SServiceMethod, eCAL::SServiceMethodInformation> client_info_map;
+  std::map<eCAL::Registration::SServiceMethod, eCAL::SServiceMethodInformation> client_info_map;
 
   // create 2 clients with the same quality of data type information
   {
@@ -113,7 +113,7 @@ TEST(core_cpp_util, ClientEqualQualities)
     eCAL::Process::SleepMS(2 * CMN_REGISTRATION_REFRESH_MS);
 
     // get all clients
-    eCAL::Util::GetClients(client_info_map);
+    eCAL::Registration::GetClients(client_info_map);
 
     // check size
     EXPECT_EQ(client_info_map.size(), 1);
@@ -122,10 +122,10 @@ TEST(core_cpp_util, ClientEqualQualities)
     std::string req_type, resp_type;
     std::string req_desc, resp_desc;
 
-    eCAL::Util::GetClientTypeNames("foo::service", "foo::method", req_type, resp_type);
+    eCAL::Registration::GetClientTypeNames("foo::service", "foo::method", req_type, resp_type);
     EXPECT_EQ(req_type,  "foo::req_type1");
     EXPECT_EQ(resp_type, "foo::resp_type1");
-    eCAL::Util::GetClientDescription("foo::service", "foo::method", req_desc, resp_desc);
+    eCAL::Registration::GetClientDescription("foo::service", "foo::method", req_desc, resp_desc);
     EXPECT_EQ(req_desc,  "foo::req_desc1");
     EXPECT_EQ(resp_desc, "foo::resp_desc1");
 
@@ -139,10 +139,10 @@ TEST(core_cpp_util, ClientEqualQualities)
     eCAL::CServiceClient client2("foo::service", { {"foo::method", service_method_info2} });
 
     // check attributes
-    eCAL::Util::GetClientTypeNames("foo::service", "foo::method", req_type, resp_type);
+    eCAL::Registration::GetClientTypeNames("foo::service", "foo::method", req_type, resp_type);
     EXPECT_EQ(req_type,  "foo::req_type1");
     EXPECT_EQ(resp_type, "foo::resp_type1");
-    eCAL::Util::GetClientDescription("foo::service", "foo::method", req_desc, resp_desc);
+    eCAL::Registration::GetClientDescription("foo::service", "foo::method", req_desc, resp_desc);
     EXPECT_EQ(req_desc,  "foo::req_desc1");
     EXPECT_EQ(resp_desc, "foo::resp_desc1");
 
@@ -153,7 +153,7 @@ TEST(core_cpp_util, ClientEqualQualities)
     eCAL::Process::SleepMS(CMN_MONITORING_TIMEOUT_MS);
 
     // get all clients again, clients should not be expired
-    eCAL::Util::GetClients(client_info_map);
+    eCAL::Registration::GetClients(client_info_map);
 
     // check size
     EXPECT_EQ(client_info_map.size(), 1);
@@ -165,10 +165,10 @@ TEST(core_cpp_util, ClientEqualQualities)
     eCAL::Process::SleepMS(2 * CMN_REGISTRATION_REFRESH_MS);
 
     // check attributes, client 1 attributes should be replaced by client 2 attributes now
-    eCAL::Util::GetClientTypeNames("foo::service", "foo::method", req_type, resp_type);
+    eCAL::Registration::GetClientTypeNames("foo::service", "foo::method", req_type, resp_type);
     EXPECT_EQ(req_type,  "foo::req_type2");
     EXPECT_EQ(resp_type, "foo::resp_type2");
-    eCAL::Util::GetClientDescription("foo::service", "foo::method", req_desc, resp_desc);
+    eCAL::Registration::GetClientDescription("foo::service", "foo::method", req_desc, resp_desc);
     EXPECT_EQ(req_desc,  "foo::req_desc2");
     EXPECT_EQ(resp_desc, "foo::resp_desc2");
   }
@@ -178,7 +178,7 @@ TEST(core_cpp_util, ClientEqualQualities)
 
   // get all clients again, all clients 
   // should be removed from the map
-  eCAL::Util::GetClients(client_info_map);
+  eCAL::Registration::GetClients(client_info_map);
 
   // check size
   EXPECT_EQ(client_info_map.size(), 0);
@@ -187,15 +187,15 @@ TEST(core_cpp_util, ClientEqualQualities)
   eCAL::Finalize();
 }
 
-TEST(core_cpp_util, ClientDifferentQualities)
+TEST(core_cpp_registration, ClientDifferentQualities)
 {
   // initialize eCAL API
-  eCAL::Initialize(0, nullptr, "core_cpp_util");
+  eCAL::Initialize(0, nullptr, "core_cpp_registration");
 
   // enable loop back communication in the same process
   eCAL::Util::EnableLoopback(true);
 
-  std::map<eCAL::Util::SServiceMethod, eCAL::SServiceMethodInformation> client_info_map;
+  std::map<eCAL::Registration::SServiceMethod, eCAL::SServiceMethodInformation> client_info_map;
 
   // create 2 clients with different qualities of data type information
   {
@@ -211,7 +211,7 @@ TEST(core_cpp_util, ClientDifferentQualities)
     eCAL::Process::SleepMS(2 * CMN_REGISTRATION_REFRESH_MS);
 
     // get all clients
-    eCAL::Util::GetClients(client_info_map);
+    eCAL::Registration::GetClients(client_info_map);
 
     // check size
     EXPECT_EQ(client_info_map.size(), 1);
@@ -220,10 +220,10 @@ TEST(core_cpp_util, ClientDifferentQualities)
     std::string req_type, resp_type;
     std::string req_desc, resp_desc;
 
-    eCAL::Util::GetClientTypeNames("foo::service", "foo::method", req_type, resp_type);
+    eCAL::Registration::GetClientTypeNames("foo::service", "foo::method", req_type, resp_type);
     EXPECT_EQ(req_type,  "foo::req_type1");
     EXPECT_EQ(resp_type, "");
-    eCAL::Util::GetClientDescription("foo::service", "foo::method", req_desc, resp_desc);
+    eCAL::Registration::GetClientDescription("foo::service", "foo::method", req_desc, resp_desc);
     EXPECT_EQ(req_desc,  "foo::req_desc1");
     EXPECT_EQ(resp_desc, "");
 
@@ -239,10 +239,10 @@ TEST(core_cpp_util, ClientDifferentQualities)
     eCAL::Process::SleepMS(2 * CMN_REGISTRATION_REFRESH_MS);
 
     // check attributes, we expect attributes from client 2 here
-    eCAL::Util::GetClientTypeNames("foo::service", "foo::method", req_type, resp_type);
+    eCAL::Registration::GetClientTypeNames("foo::service", "foo::method", req_type, resp_type);
     EXPECT_EQ(req_type,  "foo::req_type2");
     EXPECT_EQ(resp_type, "foo::resp_type2");
-    eCAL::Util::GetClientDescription("foo::service", "foo::method", req_desc, resp_desc);
+    eCAL::Registration::GetClientDescription("foo::service", "foo::method", req_desc, resp_desc);
     EXPECT_EQ(req_desc,  "foo::req_desc2");
     EXPECT_EQ(resp_desc, "foo::resp_desc2");
 
@@ -255,7 +255,7 @@ TEST(core_cpp_util, ClientDifferentQualities)
 
   // get all clients again, all clients
   // should be removed from the map
-  eCAL::Util::GetClients(client_info_map);
+  eCAL::Registration::GetClients(client_info_map);
 
   // check size
   EXPECT_EQ(client_info_map.size(), 0);
